Split ray-targeting helper into its own module and fix annotation script name
