+++ src/components/VirtualWalkthrough/xr-annotation-interaction.ts
@@ -1,61 +1,13 @@
 import { Script, Vec3, XRTYPE_VR, XrInputSource } from 'playcanvas';
+import { Annotation as PcAnnotation } from 'playcanvas/scripts/esm/annotations.mjs';
 
-export interface AnnotationHitCandidate {
-  position: Vec3;
-  radius: number;
-}
-
-/**
- * Distance along a (normalized) ray to where it first enters the sphere, or null if it never does.
- * A ray starting inside the sphere returns 0.
- */
-const raySphereEntryDistance = (origin: Vec3, dir: Vec3, center: Vec3, radius: number): number | null => {
-  const m = new Vec3().sub2(origin, center);
-  const b = m.dot(dir);
-  const c = m.dot(m) - radius * radius;
-  if (c > 0 && b > 0) {
-    return null;
-  }
-  const disc = b * b - c;
-  if (disc < 0) {
-    return null;
-  }
-  const t = -b - Math.sqrt(disc);
-
-  return t < 0 ? 0 : t;
-};
-
-/**
- * Index of the closest candidate sphere the ray enters, or null if it hits none.
- */
-export const nearestAnnotationHit = (
-  origin: Vec3,
-  direction: Vec3,
-  candidates: AnnotationHitCandidate[]
-): number | null => {
-  const dir = direction.clone().normalize();
-  let bestIndex: number | null = null;
-  let bestDistance = Infinity;
-
-  candidates.forEach((candidate, index) => {
-    const distance = raySphereEntryDistance(origin, dir, candidate.position, candidate.radius);
-    if (distance !== null && distance < bestDistance) {
-      bestDistance = distance;
-      bestIndex = index;
-    }
-  });
-
-  return bestIndex;
-};
+import { nearestAnnotationHit } from './xr-annotation-targeting';
 
 /** Local half-extent of the unit-plane hotspot quad. */
 const HOTSPOT_HALF_EXTENT = 0.5;
 
 /** Multiplier on the hotspot's world radius to make controller targeting forgiving. */
 const HIT_RADIUS_PAD = 2.5;
-
-/** PlayCanvas annotation script name. */
-const ANNOTATION_SCRIPT_NAME = 'Annotation';
 
 export class XrAnnotationInteraction extends Script {
   static scriptName = 'xrAnnotationInteraction';
@@ -86,7 +38,7 @@ export class XrAnnotationInteraction extends Script {
   private _openAnnotationUnderRay(origin: Vec3, direction: Vec3) {
     const entities = this._collectAnnotationEntities();
     const openable = entities
-      .map((ent: any) => ({ entity: ent, script: ent.script?.get(ANNOTATION_SCRIPT_NAME) }))
+      .map((ent: any) => ({ entity: ent, script: ent.script?.get(PcAnnotation.scriptName) }))
       .filter(({ script: scr }: any) => scr && scr.enabled !== false && typeof scr.onVrOpenCallback === 'function');
 
     const candidates = openable.map(({ entity: ent }: any) => ({
